Refactor to enable moving Variant interfaces to API.
