+++ core/src/main/java/org/apache/iceberg/variants/SerializedMetadata.java
@@ -113,6 +113,18 @@ class SerializedMetadata implements VariantMetadata, Variants.Serialized {
   }
 
   @Override
+  public int sizeInBytes() {
+    return buffer().remaining();
+  }
+
+  @Override
+  public int writeTo(ByteBuffer buffer, int offset) {
+    ByteBuffer value = buffer();
+    VariantUtil.writeBufferAbsolute(buffer, offset, value);
+    return value.remaining();
+  }
+
+  @Override
   public String toString() {
     return VariantMetadata.asString(this);
   }
